.requirements.txt
discord.py
pytesseract
requests